move function to outter module
